+++ spec/verikloak/rails/railtie_spec.rb
@@ -134,7 +134,6 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
   end
 
   describe '.configure_bff_guard' do
-    let(:railtie) { described_class }
     let(:middleware_stack) { double('MiddlewareStack') }
 
     before do
@@ -159,7 +158,7 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
           expect(middleware_stack).to receive(:insert_before)
             .with(::Verikloak::Middleware, ::Verikloak::BFF::HeaderGuard)
 
-          railtie.send(:configure_bff_guard, middleware_stack)
+          Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
         end
 
         it 'inserts HeaderGuard before specified middleware when bff_header_guard_insert_before is set' do
@@ -174,7 +173,7 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
           expect(middleware_stack).to receive(:insert_before)
             .with(custom_middleware, ::Verikloak::BFF::HeaderGuard)
 
-          railtie.send(:configure_bff_guard, middleware_stack)
+          Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
         end
 
         it 'inserts HeaderGuard after specified middleware when bff_header_guard_insert_after is set' do
@@ -190,7 +189,7 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
           expect(middleware_stack).to receive(:insert_after)
             .with(custom_middleware, ::Verikloak::BFF::HeaderGuard)
 
-          railtie.send(:configure_bff_guard, middleware_stack)
+          Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
         end
       end
 
@@ -205,7 +204,7 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
           expect(middleware_stack).not_to receive(:insert_before)
           expect(middleware_stack).not_to receive(:insert_after)
 
-          railtie.send(:configure_bff_guard, middleware_stack)
+          Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
         end
       end
     end
@@ -222,7 +221,163 @@ RSpec.describe Verikloak::Rails::Railtie, type: :railtie do
         expect(middleware_stack).not_to receive(:insert_before)
         expect(middleware_stack).not_to receive(:insert_after)
 
-        railtie.send(:configure_bff_guard, middleware_stack)
+        Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
+      end
+    end
+  end
+
+  describe 'BffConfigurator.configuration_valid?' do
+    let(:bff_config) { double('BffConfig') }
+
+    before do
+      stub_const('::Verikloak::BFF', Module.new)
+      allow(::Verikloak::BFF).to receive(:respond_to?).and_return(false)
+      allow(::Verikloak::BFF).to receive(:respond_to?).with(:config).and_return(true)
+      allow(::Verikloak::BFF).to receive(:respond_to?).with(:config, true).and_return(true)
+      allow(::Verikloak::BFF).to receive(:config).and_return(bff_config)
+    end
+
+    context 'when Verikloak::BFF is not defined' do
+      it 'returns true' do
+        hide_const('::Verikloak::BFF')
+        expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be true
+      end
+    end
+
+    context 'when disabled is true' do
+      it 'returns true (HeaderGuard will be inserted but internally disabled)' do
+        allow(bff_config).to receive(:respond_to?).with(:disabled).and_return(true)
+        allow(bff_config).to receive(:disabled).and_return(true)
+
+        expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be true
+      end
+    end
+
+    context 'when disabled is false or nil' do
+      before do
+        allow(bff_config).to receive(:respond_to?).with(:disabled).and_return(true)
+        allow(bff_config).to receive(:disabled).and_return(false)
+      end
+
+      context 'when trusted_proxies is not supported (legacy version)' do
+        it 'returns true' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(false)
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be true
+        end
+      end
+
+      context 'when trusted_proxies is a non-empty Array' do
+        it 'returns true' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+          allow(bff_config).to receive(:trusted_proxies).and_return(['10.0.0.0/8'])
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be true
+        end
+      end
+
+      context 'when trusted_proxies is an empty Array' do
+        it 'returns false' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+          allow(bff_config).to receive(:trusted_proxies).and_return([])
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be false
+        end
+      end
+
+      context 'when trusted_proxies is nil' do
+        it 'returns false' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+          allow(bff_config).to receive(:trusted_proxies).and_return(nil)
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be false
+        end
+      end
+
+      context 'when trusted_proxies is not an Array' do
+        it 'returns false for Set' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+          allow(bff_config).to receive(:trusted_proxies).and_return(Set.new(['10.0.0.0/8']))
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be false
+        end
+
+        it 'returns false for String' do
+          allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+          allow(bff_config).to receive(:trusted_proxies).and_return('10.0.0.0/8')
+
+          expect(Verikloak::Rails::BffConfigurator.configuration_valid?).to be false
+        end
+      end
+    end
+  end
+
+  describe '.configure_bff_guard with configuration_valid? integration' do
+    let(:middleware_stack) { double('MiddlewareStack') }
+    let(:bff_config) { double('BffConfig') }
+
+    before do
+      Verikloak::Rails.reset!
+      stub_const('::Verikloak::BFF', Module.new)
+      stub_const('::Verikloak::BFF::HeaderGuard', Class.new)
+      allow(::Verikloak::BFF).to receive(:respond_to?).and_return(false)
+      allow(::Verikloak::BFF).to receive(:respond_to?).with(:config).and_return(true)
+      allow(::Verikloak::BFF).to receive(:respond_to?).with(:config, true).and_return(true)
+      allow(::Verikloak::BFF).to receive(:config).and_return(bff_config)
+
+      Verikloak::Rails.configure do |config|
+        config.auto_insert_bff_header_guard = true
+      end
+    end
+
+    context 'when trusted_proxies is not configured' do
+      before do
+        allow(bff_config).to receive(:respond_to?).with(:disabled).and_return(true)
+        allow(bff_config).to receive(:disabled).and_return(false)
+        allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+        allow(bff_config).to receive(:trusted_proxies).and_return([])
+      end
+
+      it 'skips HeaderGuard insertion and logs warning' do
+        expect(middleware_stack).not_to receive(:insert_before)
+        expect(middleware_stack).not_to receive(:insert_after)
+
+        expect(Verikloak::Rails::RailtieLogger).to receive(:warn).with(
+          '[verikloak] Skipping BFF::HeaderGuard insertion: trusted_proxies not configured. ' \
+          'Set trusted_proxies in bff_header_guard_options to enable header validation.'
+        )
+
+        Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
+      end
+    end
+
+    context 'when disabled is true' do
+      before do
+        allow(bff_config).to receive(:respond_to?).with(:disabled).and_return(true)
+        allow(bff_config).to receive(:disabled).and_return(true)
+      end
+
+      it 'inserts HeaderGuard (internally disabled)' do
+        expect(middleware_stack).to receive(:insert_before)
+          .with(::Verikloak::Middleware, ::Verikloak::BFF::HeaderGuard)
+
+        Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
+      end
+    end
+
+    context 'when trusted_proxies is configured' do
+      before do
+        allow(bff_config).to receive(:respond_to?).with(:disabled).and_return(true)
+        allow(bff_config).to receive(:disabled).and_return(false)
+        allow(bff_config).to receive(:respond_to?).with(:trusted_proxies).and_return(true)
+        allow(bff_config).to receive(:trusted_proxies).and_return(['10.0.0.0/8'])
+      end
+
+      it 'inserts HeaderGuard' do
+        expect(middleware_stack).to receive(:insert_before)
+          .with(::Verikloak::Middleware, ::Verikloak::BFF::HeaderGuard)
+
+        Verikloak::Rails::BffConfigurator.configure_bff_guard(middleware_stack)
       end
     end
   end